api: Fail quickly if all APIs returned an error
If the internet is disconnected or things are otherwise wonky, it's
possible that all services will fail. Since we only wait for the first
successful message, this will mean that we end up waiting for the whole
amount of time indicated by --timeout, which is less than ideal. Fix
this so that we fail quickly if all APIs return an error.

Closes #47.