write_timezone: Always truncate before writing
In #93 it was reported that previous timezones were still ending up in
/etc/timezone. This sounded improbable since everything was working for
me, but then of course I checked myself and saw:

    % cat /etc/timezone
    Europe/Paris

    rk
    es

It seems like glibc only cares up to the first newline, so I never
noticed that we're not truncating properly, but we should certainly be
truncating here.

Fixes #93.